+++ src/view/Dashboard.vue
@@ -140,12 +140,13 @@ import mixin from '@/mixin'
 import finding from '@/mixin/api/finding'
 import alert from '@/mixin/api/alert'
 import iam from '@/mixin/api/iam'
+import signin from '@/mixin/api/signin'
 import StatusStatistic from '@/component/widget/statistic/StatusStatistic.vue'
 import MiniStatistic from '@/component/widget/statistic/MiniStatistic.vue'
 import CategoryStatistic from '@/component/widget/statistic/CategoryStatistic.vue'
 export default {
   name: 'PageDashboard',
-  mixins: [mixin, finding, alert, iam],
+  mixins: [mixin, finding, alert, iam, signin],
   components: {
     StatusStatistic,
     MiniStatistic,
